- streflop was disabled (likely when we switched to cmake). Turn it on for now but we need to investigate more thoroughly if we really need this?

+++ source/shared_lib/sources/map/map_preview.cpp
@@ -71,7 +71,7 @@ bool MapPreview::isCliff(int x, int y){
 			if(xToCheck < 0 || yToCheck < 0 || xToCheck >= w || yToCheck >= h){
 				//ignore
 			}
-			else if(cliffLevel <= abs(getHeight(x, y) - getHeight(xToCheck, yToCheck))){
+			else if(cliffLevel <= abs((int)(getHeight(x, y) - getHeight(xToCheck, yToCheck)))) {
 				return true;
 			}
 		}
